Check machine after create
