+++ commands/env_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/docker/machine/drivers/fakedriver"
 	"github.com/docker/machine/libmachine"
 	"github.com/docker/machine/libmachine/auth"
+	"github.com/docker/machine/libmachine/check"
 	"github.com/docker/machine/libmachine/host"
 	"github.com/docker/machine/libmachine/libmachinetest"
 	"github.com/docker/machine/libmachine/persist/persisttest"
@@ -91,7 +92,7 @@ func TestShellCfgSet(t *testing.T) {
 		description      string
 		commandLine      CommandLine
 		api              libmachine.API
-		connChecker      ConnChecker
+		connChecker      check.ConnChecker
 		noProxyVar       string
 		noProxyValue     string
 		expectedShellCfg *ShellConfig
@@ -358,7 +359,7 @@ func TestShellCfgSet(t *testing.T) {
 
 		t.Log(test.description)
 
-		defaultConnChecker = test.connChecker
+		check.DefaultConnChecker = test.connChecker
 		shellCfg, err := shellCfgSet(test.commandLine, test.api)
 		assert.Equal(t, test.expectedShellCfg, shellCfg)
 		assert.Equal(t, test.expectedErr, err)
@@ -379,7 +380,7 @@ func TestShellCfgUnset(t *testing.T) {
 		description      string
 		commandLine      CommandLine
 		api              libmachine.API
-		connChecker      ConnChecker
+		connChecker      check.ConnChecker
 		noProxyVar       string
 		noProxyValue     string
 		expectedShellCfg *ShellConfig
@@ -513,7 +514,7 @@ func TestShellCfgUnset(t *testing.T) {
 
 		t.Log(test.description)
 
-		defaultConnChecker = test.connChecker
+		check.DefaultConnChecker = test.connChecker
 		shellCfg, err := shellCfgUnset(test.commandLine, test.api)
 		assert.Equal(t, test.expectedShellCfg, shellCfg)
 		assert.Equal(t, test.expectedErr, err)
