+++ libmachine/libmachine.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/docker/machine/libmachine/auth"
 	"github.com/docker/machine/libmachine/cert"
+	"github.com/docker/machine/libmachine/check"
 	"github.com/docker/machine/libmachine/drivers"
 	"github.com/docker/machine/libmachine/engine"
 	"github.com/docker/machine/libmachine/host"
@@ -125,6 +126,14 @@ func (api *Client) Create(h *host.Host) error {
 		if err := provisioner.Provision(*h.HostOptions.SwarmOptions, *h.HostOptions.AuthOptions, *h.HostOptions.EngineOptions); err != nil {
 			return fmt.Errorf("Error running provisioning: %s", err)
 		}
+
+		// We should check the connection to docker here
+		log.Info("Checking connection to Docker...")
+		if _, _, err = check.DefaultConnChecker.Check(h, false); err != nil {
+			return fmt.Errorf("Error checking the host: %s", err)
+		}
+
+		log.Info("Docker is up and running!")
 	}
 
 	log.Debug("Reticulating splines...")
